Clean Architecture
This is an attempt to make the code more modular. All "abstract"
definitions are placed in the arena, which defines the game environment.
The concrete implementations then happen in dedicated files, e.g. in
player.rs or referee.rs.

This is based on the judgement that the arena itself is rather stable
and will not change too much. The concrete notions of Players and
Referees will change drastically. Also, there will might be a growing
number of implementations of these.