Add first incomplete MCTSPlayer
An MCTS player is added. While the implementation of the player itself
was not too hard, working out the lifetime rules was very challenging.

In the end, everything works on mutable references. The MCTS player
holds references to to base players, the arena holds references to
players and referees and everything gets borrowed.

Contrary to what was expected for some time, it is possible to mutable
borrow something that is mutably borrowed to oneself. One must ensure
not to use what is borrowed while the temporary new owner exists, but
that was not too hard here.